[REF] whatsapp: refactoring fields that uses 'free_text' field_type
The previous fix https://github.com/odoo/enterprise/commit/6f1b1a8 was not an actual solution, as the proper fix requires a view update.
This is not feasible in the stable version since we cannot force-update all customers' views.

The purpose of this task is to refactor the header_text_1, free_text_x and button_dynamic_url_x fields.
This will eliminate the need to use _origin and move the readonly attribute from the XML side to the Python file,
preventing the WhatsApp template from being recomputed upon sending.

Task-4083217

closes odoo/enterprise#68106

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>