[FIX] document: audit trail with document versionning
Normally, when using document versioning, the attachments linked to a
record are re-attached to the document when changing the version.
This is not compatible with the requirement for audit trail in Germany
(GoBD) as we want to be sure that the document won't be deleted.
This is ensure by making sure all the attachments stay attached, and
can't be deleted.

To reproduce:
* Install Accounting and Documents, as well as l10n_de
* Select/configure a German company
* Create new document folder settings for Customer Invoices
* Create a new invoice
* Send & Print
* Delete it from the attachments (it should "detach it")
* Send & Print again

opw-4095549

closes odoo/enterprise#75333

X-original-commit: 234b01a0c2607ad9a9121f7890017eb1dc857491
Related: odoo/odoo#189994
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>