[FIX] industry_fsm_sale: require company for allow_material
Problem: When the user enables allow_material on a project,
they encounter a ValidationError when adding products on a
task that has no sale order. Odoo tries to create a sale order
for the task, but there is no company_id passed into the values.

Solution: The company_id should be required when allow_material
is enabled and the company should be computed once enabled.

Steps to Reproduce on Runbot:
1. Install Sales, Field Service, Projects
2. Create a project and enable "Products on Task"
3. Create a task and assign a customer that has no sales order
4. Add products on the task and the ValidationError occurs

opw-4000580

closes odoo/enterprise#71816

X-original-commit: baf46a0dfb09f105c868c07f1fea3254e04fefdb
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Mylyna Hy (myhy) <myhy@odoo.com>