[IMP] *: assign countries to time off types
Currently, when you try to take a time off, you'll have the list of all time off types, cause there is no field "country" on the leave type. So, if you have "sick leave" in belgium, "sickness" in US and "sick holiday" in India, you'll see the three of them.

A field country_id was created on hr_leave_type model (in odoo community) to fix this issue.

This commit sets the country_id to False on the time off types defined in L10N modules. Setting the country_id on the time-off types defined in L10N modules to their corresponding countries would cause issues. That is because the time-off type records in L10Ns might be already used by allocations/leaves from multiple countries.

It also sets the country_id to False for some time-off types that are
defined outside of L10N modules to make them available for all
countries.

task-3978260

closes odoo/enterprise#68777

Related: odoo/odoo#169147
Related: odoo/upgrade#6223
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>