[FIX] documents_account: xml documents not synchronized
With a BE Company Setup
Go to Documents > Configuration > Settings
  - Enable Accounting (Centralize accounting files and documents)
  - Open Journals and synchronize the Customer Invoices journal with a workspace
Create an invoice
Click Send&Print
  - Check "BIS Billing 3.0"
Send the message, the invoice will now have 2 attachments: pdf and xml
Check the documents workspace

Issue: Only the pdf is present

This occurs because of the conditions needed to create the document.
When we send the message we create the attachment but the move does not
have attachment_ids set yet, so we fail to assign the first one.
In this phase we just need to link the xml as the pdf will be registered
when assigned as main attachment of the move

opw-4088910

closes odoo/enterprise#72884

X-original-commit: ec39016b5d6b9a93bf4a44e5437105be76f46bde
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>