[IMP] whatsapp: add phone number in account form
Purpose:
Having a Phone Number field in Whatsapp Account Configuration would allow
Support Team to test the receiving of messages a lot faster, because this is
something the clients struggle a lot with during the set-up of their account.

It's also useful for users if they want to test the incoming messages flows.

After this commit:
Added a Read-only 'Phone Number' field, which will be hidden until we
successfully connect to the API(Either from 'Sync Templates' button or
'Test Connection' button).

Task-4110445

closes odoo/enterprise#69111

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>