[FIX] stock_barcode_mrp: split incomplete to move lines
**Current behavior:**
If you have a production opened in barcode and consume part of a
component line to produce part of the final product, leave the
transfer, then re-open it and confirm -> create the backorder,
the resulting backorder will have split the remaining raw moves
in an unintuitive manner.

**Expected behavior:**
The backorder should have one move for the remaining component
quantity.

**Steps to reproduce:**
1. Create a production for 10 of some final product consuming 10
     some component -> Confirm

2. Open the production in barcode, add 5 of the final product to
     its line and 5 of the component to its line -> exit the
     transfer view

3. Reopen the production and validate it -> create backorder

4. Open the backorder to see the odd split of the component
     product's moves

**Cause of the issue:**
When creating the backorder we normally expect each component
product to be encapsulated by a single line- so the split that
occurs when we leave the transfer initially without validating
which creates 2 moves for the same product means we get 2 moves
for the component for half of the remaining quantity for that
component.

**Fix:**
Incomplete barcode lines for production transfers should get
split into additional move lines as opposed to moves.

opw-4148050

closes odoo/enterprise#71776

X-original-commit: 4ad9617cc40549982f85346775713f1560270199
Related: odoo/odoo#183364
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Vincent Ethan <etvi@odoo.com>