[FIX] sign: automatic cancel of sign requests in emails
This commit aims to fix the following issue:
When we send a sign request, an email is sent containing some link
for canceling the sign request, that is when clicked
the sign request is canceled instantly.

The issue is that some spam detectors is going to click every
link in the email while validating it. Resulting in the sign
request being canceled without the partner knowing about it.

The commit changes the action of the email cancel link to
displaying a form and asking the user to submit it to
confirm the sign request cancel.

Task: 4149784
X-original-commit: e08ab3deaa4a6b30b6fb2e72ef4458daa0c3c209
Part-of: odoo/enterprise#71692
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>