[IMP] appointment,*:  create dynamic slot intervals
*= pos_restaurant_appointment,appointment_hr_recruitment

Purpose
========
Currently, offering multiple arrival times
(on the hour, or at 15, 30, and 45 minutes past) creates excessive slot entries
and is not intuitive for users. The goal is to streamline this process and make
the options clearer and easier to manage.

Specification
==============
Create a new field "appointment_slot_creation" with a default value 1.0
which decides when to start creating slots, based on the start time.

1).If the appointment_slot_creation is < duration it will create
multiple schedules.
EG. If the slot time is 9 - 12 with a duration of 1 hr and
appointment_slot_creation is set as 15 min then it will create slots
like (9:00, 9:15, 9:30, 9:45, 10:00, 10:15, 10:30, 10:45, 11:00).

2).If the field value is > duration it will allow setting up buffer time
between slots.
EG. If the slot time is 9-12 with the duration of 1 hr and
appointment_slot_creation is set as 1:15 min then it will
create slots like (9:00, 10:15).

3).Some demo data are also change.

4).If appointment type is custom then we will show the field in the form view
but instead of it's input box we will show the static string like

Duration  -   Depending on the slots
Create a slot every - Schedule line

Upgrade PR-https://github.com/odoo/upgrade/pull/6626﻿﻿

Task-4143918

closes odoo/enterprise#70098

Related: odoo/upgrade#6626
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>