[IMP] planning: rework splitting a weekend planning slot
Prior to this commit, a planning slot spanning over a weekend would not
get the correct start/end time, since it will take the availabilities
of the resource into account.

As we do not take working calendars as reference for flexible
resource. For example, for a Sunday 8AM to Monday 5PM shift, the expected
split is:
- Sunday 8:00 to Sunday 23:59
- Monday 00:00 to Monday 17:00

This commit reworks the split of a weekend slot as follows:
- open slots will also follow the same split as the one for flexible resources.
- weekend slots for regular resources will have their start/end time
  set to 8AM/5PM instead of having empty slots of 0 minutes. This will allow
  to pass shifts to other flexible resources more easiy.
- JS Tour test added for the split of a weekend slot.
- reduce the number of ORM calls by providing a specific method for the split.

closes odoo/enterprise#69963

Task-id: 4074987C
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>