[IMP] website_sale_*: misc improvements on product page
Variant Mixin logic is only used on ecommerce since the drop of the old
product configurator (optional products modal) replaced by the backend one,
thus `isWebsite` is always truthy and useless to check

Avoid verification of rental dates when the product is not a rental product.

Also adapt to community changes with the removal of `price_extra` key from
combination info, and the removal of the comparison price when not enabled.

closes odoo/enterprise#70313

Related: odoo/odoo#180774
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>