[ADD] l10n_br_{edi_pos,avatax}: Brazilian EDI in the POS with Avatax
This module implements issueing NFC-e receipts through the POS. The
process is conceptually similar to what we do for invoices. First
taxes are calculated, and then we e-invoice.

The calculated taxes will never change the order total. NFC-e mandates
taxes to always be included in the price so we don't need any
additional RPC call before payment.

One way of achieving this was through l10n_br_edi, forcing every
pos.order to be invoiced and then following the implemented flows on
account.move. We decided against it because:

- It leads to a large amount of mostly unnecessary invoices,
- It's conceptually strange to the user, NFC-e "invoices" resemble POS
receipts more than they do invoices,
- The flow in the POS is simpler, we handle tax calculation and EDI in
one atomic step.

We therefore chose to re-implement EDI for pos.order. The downside of
this approach is that we temporarily need to copy some code from
l10n_br_edi. In master this code can be consolidated in a common
mixin.

The integration tries to never block POS sales. You're allowed to
retry EDI later. When EDI fails, the POS user is informed and we fall
back to the standard receipt (marked as a "receipt without fiscal
value").

For refunds we still go through account.move, as NFC-e doesn't support
refunds (must be NF-e, which is what we already support for
account.move).

Notes about forward-port to 18.1:

A small bug was fixed in l10n_br_avatax:
_compute_l10n_br_avatax_warnings() referenced self in a loop which
broke the POS. Before this, it seems the field was never computed on
multiple records.

_get_available_product_domain() no longer exists. It was replaced by a
_load_pos_data_domain() override on product.template. This new
override has an extra part that makes sure to exclude combo products
that have combo items with price-excluded taxes. _load_pos_data() of
product.template will load combo products indiscriminately of the
provided domain, thus bypassing our goal of not loading products with
price-excluded taxes.

JS method names were camelCase'd, to be more similar with the POS.

In this version, runbot detected a string we forgot to translate. It's
now translated.

export_for_printing() no longer exists and was removed. The QWeb
overrides now access members directly on the POS order where
possible. We use the same methods as before when something needs to be
computed, but they've been rewritten to use `this` instead of a
parameter.

A bug was fixed in the automatic adjustment entry. It was assumed that
_prepare_aml_values_list_per_nature() would group lines per income
account like in the closing entry, but this is not the case. One line
per product is generated instead. A test was added to to cover this
case: test_05_edi_after_session_closed_different_product_same_account

task-3564171

closes odoo/enterprise#82039

X-original-commit: 02075a25c6616cfe8df1262e29bbabb4d968b774
Related: odoo/odoo#203199
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Joren Van Onder (jov) <jov@odoo.com>