[REF] pos_self_order_iot: remove open kiosk popup
When clicking on "Start" on a self order PoS, a popup opened with the link to the self order page.
We removed this modal, and requested kiosk opening on the iot box via websocket for simplicity reasons.

In addition, a Kiosk PoS is automatically set to work with the IoT Box if selected in the autoconfigure modal when
pairing a new IoT Box.

Community PR: [https://github.com/odoo/odoo/pull/179767](https://github.com/odoo/odoo/pull/179767)

closes odoo/enterprise#69771

Task: 4166305
Related: odoo/odoo#179767
Related: odoo/upgrade#7554
Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>