[FIX] sale_purchase[stock]_inter_company_rules: unspecified multi-currency dependency
Both modules have tests with an unstated dependency on the
multi-currency mode, as they unconditionally try setting currencies
into forms.

They will pass if they run in a multi-currency-enabled database[^1]
but if the context is non-multi-currency (e.g. no l10n module is
explicitly installed so only `l10n_us` is present and USD is the only
currency) then they fail with

    can't write on invisible field 'currency_id'

[^1]: generally the case on runbot where `l10n_be` is ~always
    installed, which enables EUR, which being a second currency
    automatically enables multi-currency

closes odoo/enterprise#72354

Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>