[FIX] mrp_mps: show underreplenished products as 'to replenish'
The existing filter searches only for products to launch (not yet
replenished) if searched for products to replenish, omitting the
products to relaunch (replenished insufficiently), which are displayed
only with the 'replenishment too low' filter. This commit ensures
the products to relaunch appear under both filters.

Task ID: 4104208

Part-of: odoo/enterprise#68419
Related: odoo/odoo#176892
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>