[FIX] planning: allocated hours not recomputed in shift duplication
Steps to reproduce:
-------------------
1. Install Planning app
2. Have an employee A with 40h/week working schedule and an employee B with 35h/week working schedule
3. Create a shift and assign it to employee A, the default allocated hours should be 8h
4. From the Gantt view, duplicate this shift and assign it to employee B (use drag and drop)
5. Problem: The allocated hours of the duplicated shift are not recomputed, they are equal to 8h instead of 7h

Fix:
-------------------
When duplicating a shift and assigning it to another resource, the copy() method is called with the resource_id
of the target resource (an employee in this case). But _compute_allocated_hours() is not called so we have to
call it explicitly in this case.

version-17.0
task-3978590

closes odoo/enterprise#73432

X-original-commit: 7394db9928f96ed6449c32cb8de74f92cc60333c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>