[IMP] sale_timesheet_enterprise: show billable time target fields in list view
Before this commit, when the user would like to set billable time
targets to his employees and clicks on `Set Employee Billable Time
Targets` action displayed in `Timesheets > Configuration > Settings`
when `Billing Rate Indicators` is enabled. The user will be redirected
to the kanban view of employees but he does not really know which field
he has to set if he tries the feature, and he has to go to the form view
and find the field and set it for each employee using timesheets app.

This commit reviews the action loaded in that button to directly show
the list view with the fields needed. By doing that, the user will
directly see the field to set for his employees and he will be able to
mass edit thanks to the list view.

task-4179246

closes odoo/enterprise#69914

Related: odoo/odoo#186559
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>