[IMP] sign: make redirect button label mandatory
This commits makes the redirect button label mandatory when there is
a specified redirect URL. It also improves the hiding of the redirect
label field when there is no redirect URL specified. After this commit,
we changed the default redirect url text from 'Open Link' to 'Close'.
Additionally, when the redirect URL is specified, we no longer show
the Close button for simplifying the UI and not repeat the redirect behavior.

task-4174120

closes odoo/enterprise#74195

X-original-commit: ed8ab688cc970ef704f52431654c8fc85e476dee
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>