[FIX] sale_purchase_inter_company_rules: allow interco dropship
Fixes two uses cases:
- Sale Order in Company A that dropships goods *stored* in Company B to
the Customer.
- Sale Order in Company A that dropships goods from company B to the
Customer. But in company B, we also dropship from the vendor to the
customer.

In both cases, make sure to properly send goods to the Inter-Company
Transit location for inter-co transactions, so that goods (and
lot/serials) can be properly linked.

opw-4163612

closes odoo/enterprise#71181

X-original-commit: 70dd97c6d79e4f3bf61512e122c80c2a3f3f6697
Related: odoo/odoo#182294
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>