[FIX] mrp_account_enterprise: fix expected_component_cost_unit
Steps to reproduce the issue:

1. Install the MRP module
2. Go to Production Analysis
3. Click on the pivot table
4. Select the `Expected component cost / unit` field
5. The miscalculation will be show for
`expected_component_cost_unit`. It is dividing by the
MO quantity produced rather than the BOM quantity.

Explanation:

The issue lies in the SQL query, where the expected
cost is divided by `prod_qty.product_quantity`, meaning
we are using the quantity produced by a MO. instead, it
should be divided by the BOM quantity.

Fix reasoning:

With this commit, the issue is resolved by dividing the
expected cost by the BOM `product_qty`.

closes odoo/enterprise#72721

X-original-commit: d2cadefe8a4b1f4b88b0e7cfb47ddb465621a3b7
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: José Carlos da Silva Cruz (jcsc) <jcsc@odoo.com>