[IMP] mrp_workorder: show 'Pending' in place of timer if WO pending or waiting
The timer field is normally left empty if the WO is in 'ready' or
'pending' states, and the timer appears only when it's in progress or
finished. This commit adds the text 'Pending' in the place of the
timer if the WO is in pending state, and replaces the blank timer with
it if it's in waiting stage.

Task ID: 4154879

Part-of: odoo/enterprise#71679
Related: odoo/odoo#183175
Signed-off-by: William Henrotin (whe) <whe@odoo.com>