[IMP] account_intrastat,l10n_be_intrastat: Add support for service
Currently, the Intrastat report is restricted to products whose product type is different than services.
Though, statistics report can also be required on Services, such as Declarations F01DGS & F02CMS

(Manual : https://www.nbb.be/doc/dq/e_pdf_bb/f01dgs-f02cms_manual_en.pdf)

This commit intend to improve the Intrastat report to also include services. Since commodities and services share a lot of similarities, we didn't create a new report but rather added a filter inside the current Intrastat report. It also adds a new column for the service categories and all the service codes.

On the product, the string of the commodity code changes if the product is a commodity or a service (and become Service Code in that case)

task-2835786

closes odoo/enterprise#70091

Signed-off-by: John Laterre (jol) <jol@odoo.com>