[IMP] planning: remove and rename labels in planning slot list view
Before this commit, the label displayed for planned dates in the list
view of `planning.slot` model is `Date Start` which is not really good
since the column displayed the date start but also the date end thanks
to the daterange widget. Moreover, the `allocated_hours` displayed in
the list view is a bit too far of the planned dates because of the
column name since the name is much longer than the values on each cell
inside that column.

This commit renames the label for the column contained the planned dates
into `Planned Dates`, the allocated hours value is added in muted in the
planned dates column when the row is in readonly since that information
makes more sense with the planned dates data and moreover, the column
containing the allocated hours was not editable.

task-4034478

closes odoo/enterprise#69974

Related: odoo/odoo#180100
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>