[FIX] web_studio: allow click on hidden buttons inside button box
Steps to reproduce
==================

- Go to contacts
- Open any record
- Open studio
- Click on the "More" button at the top right next to other buttons
- A dropdown is displayed containing more buttons
- Try clicking on one of them

=> Nothing happens

Cause of the issue
==================

The dropdown is outside `this.viewRenderer.el`.
Clicks on it are ignored. This also causes studio specific styling to
not be applied.

Solution
========

The ButtonBox component from web displays maximum number of buttons
outside the dropdown. Then remaining buttons are displayed inside the
dropdown.

We can copy this approach, but instead of rendering extra buttons inside
the dropdown, we can show them on the following line.

opw-4164661

closes odoo/enterprise#71079

X-original-commit: 9d3c95b6a7282e9850b7027d3fefbf280c3c2f1c
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>