[FIX] whatsapp: allow receiving large file from WhatsApp
Currently a below occurs or content is not receiving to odoo WhatsApp
when the user uploads a large file (tried with > 35 MB).

error: `MemoryError: null`

At line [1], the code `'error' in res.json()` is used, which reads all the
json content of `res`, but at times users upload large files, it will cause
the above error because the `json()` tries to read all bytes from res, which
is a very large amount to handle in memory.

This commit will fix the above issue by returning a response if the response
contains content_length more than 10 MB.

[1] - https://github.com/odoo/enterprise/blob/aca7ae2a7cf4aad5427a60d5cad60d08774357d0/whatsapp/tools/whatsapp_api.py#L46

sentry-5810101850

closes odoo/enterprise#91985

X-original-commit: 59beae420a7a5f8f4028a10164cc6acab0b6a9e0
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: ANSARI MAHAMADASIF (maan) <maan@odoo.com>