[FIX] industry_fsm_sale: fix invoice status of a SO with zero price product
Before this commit:
When Anglo-Saxon accounting is enabled and a product with sale price of zero is
added to SO through field service, the invoice_status show 'Nothing to Invoice',
can not create an invoice for a zero priced product associated with an FSM task.

After this commit:
With Anglo-Saxon accounting enabled, adding product with a zero sale price to an
SO through field service will display the invoice_status as 'To Invoice' instead
of 'Nothing to Invoice', it is now possible to create an invoice for a zero
priced product associated with an FSM task when Anglo-Saxon is enabled.

task-3957962

closes odoo/enterprise#91247

X-original-commit: d20dbc656787efddb2695ff06ad9aeea7117b315
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>