[IMP] website_{*}: pickup location ui/ux improvements
*: delivery_sendcloud, sale_fedex

This commit introduces minor improvements to the pickup locations
feature available in FedEx and Sendcloud modules. The unit of
measurement of the radius is now readonly and set to be KM or mile
depending on the measurement system defined by the user. The default
value of the radius is also set to be 10 KMs (or miles). The label of
the field is changed to be more concise.

Task-4173210

X-original-commit: 3870b0e89375bdb18cc67eea2a9882ef141f7446
Part-of: odoo/enterprise#71203
Related: odoo/odoo#182331
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Mohammad Abdulmoneim (abdu) <abdu@odoo.com>